create setupAvailability screen & replace in paginator create index file for availability subcomponents
