+++ src/screens/Onboarding/Onboarding.tsx
@@ -2,22 +2,17 @@ import { Paginator } from 'components/Paginator/Paginator'
 import { SetUpProfile } from './SetUpProfile'
 import { Role } from './Role'
 import { WhatsNext } from './WhatsNext'
-import { Availability } from 'components/Availability/Availability'
 import './Onboarding.scss'
-import { useState } from 'react'
-import { defaultAvailability } from 'utils/data/userConstants'
+import { SetupAvailability } from './SetupAvailability'
 
 export const Onboarding = () => {
-  const [days, setDays] = useState(defaultAvailability)
-
   const orderedPages = [
     {
       component: Role,
       title: 'Role',
     },
     {
-      component: Availability,
-      props: { days, setDays },
+      component: SetupAvailability,
       title: 'Availability',
     },
     {
